+++ src/main/java/com/lapetitesyrienne/api/models/Article.java
@@ -20,13 +20,13 @@ public abstract class Article {
     @NotBlank
     private String description;
     private String image;
-    private float price;
+    private Double price;
     private Date createdAt;
 
     public Article() {
     }
 
-    public Article(String name, String description, String image, float price) {
+    public Article(String name, String description, String image, Double price) {
         this.name = name;
         this.description = description;
         this.image = image;
@@ -66,11 +66,11 @@ public abstract class Article {
         this.image = image;
     }
 
-    public float getPrice() {
+    public Double getPrice() {
         return price;
     }
 
-    public void setPrice(float price) {
+    public void setPrice(Double price) {
         this.price = price;
     }
